Refactor by exbedding Blacklist and DigestChoice.
