+++ src/main/java/net/shibboleth/tool/xmlsectool/XmlSecToolCommandLineArguments.java
@@ -20,12 +20,7 @@ package net.shibboleth.tool.xmlsectool;
 import java.io.PrintStream;
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.Collections;
 import java.util.List;
-import java.util.Set;
-import java.util.TreeSet;
-
-import org.opensaml.xml.signature.SignatureConstants;
 
 import com.beust.jcommander.JCommander;
 import com.beust.jcommander.Parameter;
@@ -34,253 +29,6 @@ import com.beust.jcommander.ParameterException;
 /** Command line arguments for the {@link XmlSecTool} command line tool. */
 public class XmlSecToolCommandLineArguments {
 
-    /**
-     * A blacklist of digest and signature algorithms we should not accept during
-     * signature verification.
-     */
-    public class Blacklist {
-        
-        /**
-         * Ordered set of blacklisted digest algorithm URIs.
-         */
-        private final Set<String> digestBlacklist = new TreeSet<String>();
-        
-        /**
-         * Ordered set of blacklisted signature algorithm URIs.
-         */
-        private final Set<String> signatureBlacklist = new TreeSet<String>();
-        
-        /**
-         * Constructor.
-         *
-         * Initializes the blacklist with those algorithms that should be
-         * blacklisted by default.
-         */
-        public Blacklist() {
-            // MD5
-            addDigestAlgorithm(SignatureConstants.ALGO_ID_DIGEST_NOT_RECOMMENDED_MD5);
-            addSignatureAlgorithm(SignatureConstants.ALGO_ID_SIGNATURE_NOT_RECOMMENDED_RSA_MD5);
-            
-            // SHA-1
-            addDigest(DigestChoice.SHA1);
-        }
-        
-        /**
-         * Blacklist an individual digest algorithm.
-         * 
-         * @param uri algorithm URI to blacklist
-         */
-        private void addDigestAlgorithm(String uri) {
-            digestBlacklist.add(uri);
-        }
-        
-        /**
-         * Blacklist an individual signature algorithm.
-         * 
-         * @param uri algorithm URI to blacklist
-         */
-        private void addSignatureAlgorithm(String uri) {
-            signatureBlacklist.add(uri);
-        }
-        
-        /**
-         * Blacklist the digest and signature algorithms associated with
-         * a {@link DigestChoice}.
-         * 
-         * @param digestChoice {@link DigestChoice} to add to blacklist
-         */
-        public void addDigest(DigestChoice digestChoice) {
-            addDigestAlgorithm(digestChoice.getDigestAlgorithm());
-            addSignatureAlgorithm(digestChoice.getRsaAlgorithm());
-            addSignatureAlgorithm(digestChoice.getEcdsaAlgorithm());
-        }
-        
-        /**
-         * Returns <code>true</code> if the indicated algorithm URI is blacklisted for
-         * use as a digest algorithm.
-         * 
-         * @param alg digest algorithm URI to check
-         * @return <code>true</code> if the algorithm is blacklisted
-         */
-        public boolean isBlacklistedDigest(String alg) {
-            return digestBlacklist.contains(alg);
-        }
-        
-        /**
-         * Returns <code>true</code> if the indicated algorithm URI is blacklisted for
-         * use as a signature algorithm.
-         * 
-         * @param alg signature algorithm URI to check
-         * @return <code>true</code> if the algorithm is blacklisted
-         */
-        public boolean isBlacklistedSignature(String alg) {
-            return signatureBlacklist.contains(alg);
-        }
-        
-        /**
-         * Returns an unmodifiable view on the set of blacklisted digest algorithms.
-         * 
-         * @return set of blacklisted algorithms
-         */
-        public Collection<String> getDigestBlacklist() {
-            return Collections.unmodifiableCollection(digestBlacklist);
-        }
-        
-        /**
-         * Returns an unmodifiable view on the set of blacklisted signature algorithms.
-         * 
-         * @return set of blacklisted algorithms
-         */
-        public Collection<String> getSignatureBlacklist() {
-            return Collections.unmodifiableCollection(signatureBlacklist);
-        }
-        
-        /**
-         * Empties the digest and signature blacklists.
-         */
-        public void clear() {
-            digestBlacklist.clear();
-            signatureBlacklist.clear();
-        }
-    }
-    
-    /**
-     * The digest method to use in the various signature algorithms.
-     */
-    public static enum DigestChoice {
-
-        /**
-         * SHA-1 digest.
-         */
-        SHA1("SHA-1",
-                SignatureConstants.ALGO_ID_DIGEST_SHA1,
-                SignatureConstants.ALGO_ID_SIGNATURE_RSA_SHA1,
-                SignatureConstants.ALGO_ID_SIGNATURE_ECDSA_SHA1),
-        
-        /**
-         * SHA-256 digest.
-         */
-        SHA256("SHA-256",
-                SignatureConstants.ALGO_ID_DIGEST_SHA256,
-                SignatureConstants.ALGO_ID_SIGNATURE_RSA_SHA256,
-                SignatureConstants.ALGO_ID_SIGNATURE_ECDSA_SHA256),
-        
-        /**
-         * SHA-384 digest.
-         */
-        SHA384("SHA-384",
-                SignatureConstants.ALGO_ID_DIGEST_SHA384,
-                SignatureConstants.ALGO_ID_SIGNATURE_RSA_SHA384,
-                SignatureConstants.ALGO_ID_SIGNATURE_ECDSA_SHA384),
-        
-        /**
-         * SHA-512 digest.
-         */
-        SHA512("SHA-512",
-                SignatureConstants.ALGO_ID_DIGEST_SHA512,
-                SignatureConstants.ALGO_ID_SIGNATURE_RSA_SHA512,
-                SignatureConstants.ALGO_ID_SIGNATURE_ECDSA_SHA512);
-        
-        /**
-         * Other name (with hyphens, etc.) used as an alternative to the enum name.
-         */
-        private final String otherName;
-        
-        /**
-         * Digest algorithm.
-         */
-        private final String digestAlgorithm;
-        
-        /**
-         * RSA signature algorithm.
-         */
-        private final String rsaAlgorithm;
-        
-        /**
-         * ECDSA signature algorithm.
-         */
-        private final String ecdsaAlgorithm;
-        
-        /**
-         * Constructor.
-         * 
-         * @param otherNameArg an alternative name for the enum.
-         * @param digestArg digest algorithm URI
-         * @param rsaArg RSA signature algorithm URI
-         * @param ecdsaArg ECDSA signature algorithm URI
-         */
-        private DigestChoice(final String otherNameArg,
-                final String digestArg, final String rsaArg, final String ecdsaArg) {
-            otherName = otherNameArg;
-            digestAlgorithm = digestArg;
-            rsaAlgorithm = rsaArg;
-            ecdsaAlgorithm = ecdsaArg;
-        }
-        
-        /**
-         * Returns the digest algorithm URI for this digest choice.
-         * 
-         * @return algorithm URI
-         */
-        public String getDigestAlgorithm() {
-            return digestAlgorithm;
-        }
-        
-        /**
-         * Returns the RSA signature algorithm URI for this digest choice.
-         * 
-         * @return algorithm URI
-         */
-        public String getRsaAlgorithm() {
-            return rsaAlgorithm;
-        }
-        
-        /**
-         * Returns the ECDSA signature algorithm URI for this digest choice.
-         * 
-         * @return algorithm URI
-         */
-        public String getEcdsaAlgorithm() {
-            return ecdsaAlgorithm;
-        }
-        
-        /**
-         * Indicates whether the enum can be called by the provided name.
-         * 
-         * The name is compared ignoring case against the enum name and
-         * against the "other" name.
-         * 
-         * @param name name to check against
-         * @return <code>true</code> if and only if the enum can be called by the provided name
-         */
-        public boolean hasName(final String name) {
-            if (name.equalsIgnoreCase(name())) {
-                return true;
-            }
-            if (name.equalsIgnoreCase(otherName)) {
-                return true;
-            }
-            return false;
-        }
-
-        /**
-         * Finds the {@link DigestChoice} for a given digest name.
-         * 
-         * @param name name of the digest to be found
-         * 
-         * @return {@link DigestChoice} represented by the name
-         */
-        public static DigestChoice find(String name) {
-            for (DigestChoice choice: values()) {
-                if (choice.hasName(name)) {
-                    return choice;
-                }
-            }
-            return null;
-        }
-
-    }
-    
     /*
      * Checkstyle: JavadocVariable OFF
      * Checkstyle: JavadocMethod OFF
